split files as per suggestions
