+++ tests/integration/tasks/test_orchestrator.py
@@ -160,7 +160,7 @@ def test_manage_not_start(
     )
 
     with mock.patch(
-        "aap_eda.services.activation.manager.get_current_job",
+        "aap_eda.services.activation.activation_manager.get_current_job",
         return_value=job_mock,
     ):
         orchestrator._manage(ProcessParentType.ACTIVATION, activation.id)
@@ -277,7 +277,7 @@ def test_max_running_activation_after_start_job(
         ProcessParentType.ACTIVATION, activation.id, ActivationRequest.START
     )
     with mock.patch(
-        "aap_eda.services.activation.manager.get_current_job",
+        "aap_eda.services.activation.activation_manager.get_current_job",
         return_value=job_mock,
     ):
         orchestrator._manage(ProcessParentType.ACTIVATION, activation.id)
